i#1569 AArch64: Implement trace optimisation.
The optimisation is not yet enabled by default, though some programs
run faster, and correctly, with "-enable_traces".

This is based on work done by Kevin Zhou in July 2016.

+++ core/arch/instr.h
@@ -295,7 +295,7 @@ typedef enum _dr_pred_type_t {
 #define PREFIX_PRED_BITS 5
 #define PREFIX_PRED_BITPOS (32 - PREFIX_PRED_BITS)
 #define PREFIX_PRED_MASK \
-    (((1 << PREFIX_PRED_BITS)-1) << PREFIX_PRED_BITPOS) /*0xf8000000 */
+    ((((uint)1 << PREFIX_PRED_BITS)-1) << PREFIX_PRED_BITPOS) /*0xf8000000 */
 /* DR_API EXPORT BEGIN */
 
 /**
